Stub out modules, move status checkers.
