+++ attest/verifier/src/ias.rs
@@ -1,0 +1,5 @@
+// Copyright (c) 2018-2021 The MobileCoin Foundation
+
+//! Verifiers which operate on the contents of the
+//! [`VerificationReportData`](::mc_attest_core::VerificationReportData)
+//! structure.
