+++ attest/verifier/src/lib.rs
@@ -10,7 +10,14 @@
 #![doc = include_str!("../README.md")]
 #![no_std]
 
+mod ias;
+mod quote;
+mod report_body;
+mod status;
+
 extern crate alloc;
+
+pub use crate::status::{MrEnclaveVerifier, MrSignerVerifier};
 
 cfg_if::cfg_if! {
     if #[cfg(feature = "sgx-sim")] {
@@ -653,196 +660,6 @@ impl Verify<VerificationReportData> for NonceVerifier {
     }
 }
 
-/// A helper function used to check exceptions to the quote error = fail rule.
-fn check_ids(quote_status: &IasQuoteResult, config_ids: &[String], sw_ids: &[String]) -> bool {
-    match quote_status {
-        Ok(_) => true,
-        Err(IasQuoteError::ConfigurationNeeded { advisory_ids, .. }) => {
-            advisory_ids.iter().all(|id| config_ids.contains(id))
-        }
-        Err(IasQuoteError::SwHardeningNeeded { advisory_ids, .. }) => {
-            advisory_ids.iter().all(|id| sw_ids.contains(id))
-        }
-        Err(IasQuoteError::ConfigurationAndSwHardeningNeeded { advisory_ids, .. }) => advisory_ids
-            .iter()
-            .all(|id| config_ids.contains(id) || sw_ids.contains(id)),
-        Err(_) => false,
-    }
-}
-
-/// A [`VerifyIasReportData`] implementation that will check if the enclave in
-/// question has the given MrEnclave, and has no other IAS report status issues.
-#[derive(Clone, Debug, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
-pub struct MrEnclaveVerifier {
-    mr_enclave: MrEnclave,
-    config_ids: Vec<String>,
-    sw_ids: Vec<String>,
-}
-
-impl MrEnclaveVerifier {
-    /// Create a new status verifier that will check for the existence of the
-    /// given MrEnclave.
-    pub fn new(mr_enclave: MrEnclave) -> MrEnclaveVerifier {
-        Self {
-            mr_enclave,
-            config_ids: Default::default(),
-            sw_ids: Default::default(),
-        }
-    }
-
-    /// Assume an enclave with the specified measurement does not need
-    /// BIOS configuration changes to address the provided advisory ID.
-    ///
-    /// This method should only be used when advised by an enclave author.
-    pub fn allow_config_advisory(&mut self, id: &str) -> &mut Self {
-        self.config_ids.push(id.to_owned());
-        self
-    }
-
-    /// Assume an enclave with the specified measurement does not need
-    /// BIOS configuration changes to address the provided advisory IDs.
-    ///
-    /// This method should only be used when advised by an enclave author.
-    pub fn allow_config_advisories(&mut self, ids: &[&str]) -> &mut Self {
-        for id in ids {
-            self.config_ids.push((*id).to_owned());
-        }
-        self
-    }
-
-    /// Assume the given MrEnclave value has the appropriate software/build-time
-    /// hardening for the given advisory ID.
-    ///
-    /// This method should only be used when advised by an enclave author.
-    pub fn allow_hardening_advisory(&mut self, id: &str) -> &mut Self {
-        self.sw_ids.push(id.to_owned());
-        self
-    }
-
-    /// Assume the given MrEnclave value has the appropriate software/build-time
-    /// hardening for the given advisory IDs.
-    ///
-    /// This method should only be used when advised by an enclave author.
-    pub fn allow_hardening_advisories(&mut self, ids: &[&str]) -> &mut Self {
-        for id in ids {
-            self.sw_ids.push((*id).to_owned());
-        }
-        self
-    }
-}
-
-impl From<Signature> for MrEnclaveVerifier {
-    fn from(src: Signature) -> Self {
-        Self::new(MrEnclave::from(*(src.mrenclave())))
-    }
-}
-
-impl Verify<VerificationReportData> for MrEnclaveVerifier {
-    fn verify(&self, data: &VerificationReportData) -> bool {
-        if let Ok(report_body) = data.quote.report_body() {
-            self.mr_enclave == report_body.mr_enclave()
-                && check_ids(&data.quote_status, &self.config_ids, &self.sw_ids)
-        } else {
-            false
-        }
-    }
-}
-
-/// A [`VerifyIasReportData`] implementation that will check if the enclave in
-/// question has the given MrSigner value, and has no other IAS report status
-/// issues.
-#[derive(Clone, Debug, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
-pub struct MrSignerVerifier {
-    mr_signer: MrSigner,
-    product_id: ProductId,
-    minimum_svn: SecurityVersion,
-    config_ids: Vec<String>,
-    sw_ids: Vec<String>,
-}
-
-impl MrSignerVerifier {
-    /// Create a new status verifier that will check for the existence of the
-    /// given MrSigner.
-    pub fn new(
-        mr_signer: MrSigner,
-        product_id: ProductId,
-        minimum_svn: SecurityVersion,
-    ) -> MrSignerVerifier {
-        Self {
-            mr_signer,
-            product_id,
-            minimum_svn,
-            config_ids: Default::default(),
-            sw_ids: Default::default(),
-        }
-    }
-
-    /// Assume an enclave with the specified measurement does not need
-    /// BIOS configuration changes to address the provided advisory ID.
-    ///
-    /// This method should only be used when advised by an enclave author.
-    pub fn allow_config_advisory(&mut self, id: &str) -> &mut Self {
-        self.config_ids.push(id.to_owned());
-        self
-    }
-
-    /// Assume an enclave with the specified measurement does not need
-    /// BIOS configuration changes to address the provided advisory IDs.
-    ///
-    /// This method should only be used when advised by an enclave author.
-    pub fn allow_config_advisories(&mut self, ids: &[&str]) -> &mut Self {
-        for id in ids {
-            self.config_ids.push((*id).to_owned());
-        }
-        self
-    }
-
-    /// Assume an enclave with the specified measurement has the appropriate
-    /// software/build-time hardening for the given advisory ID.
-    ///
-    /// This method should only be used when advised by an enclave author.
-    pub fn allow_hardening_advisory(&mut self, id: &str) -> &mut Self {
-        self.sw_ids.push(id.to_owned());
-        self
-    }
-
-    /// Assume an enclave with the specified measurement has the appropriate
-    /// software/build-time hardening for the given advisory IDs.
-    ///
-    /// This method should only be used when advised by an enclave author.
-    pub fn allow_hardening_advisories(&mut self, ids: &[&str]) -> &mut Self {
-        for id in ids {
-            self.sw_ids.push((*id).to_owned());
-        }
-        self
-    }
-}
-
-impl From<Signature> for MrSignerVerifier {
-    fn from(src: Signature) -> Self {
-        Self::new(src.mrsigner().into(), src.product_id(), src.version())
-    }
-}
-
-impl From<&Signature> for MrSignerVerifier {
-    fn from(src: &Signature) -> Self {
-        Self::new(src.mrsigner().into(), src.product_id(), src.version())
-    }
-}
-
-impl Verify<VerificationReportData> for MrSignerVerifier {
-    fn verify(&self, data: &VerificationReportData) -> bool {
-        if let Ok(report_body) = data.quote.report_body() {
-            self.mr_signer == report_body.mr_signer()
-                && report_body.product_id() == self.product_id
-                && report_body.security_version() >= self.minimum_svn
-                && check_ids(&data.quote_status, &self.config_ids, &self.sw_ids)
-        } else {
-            false
-        }
-    }
-}
-
 /// A [`VerifyIasReportData`] implementation which applies a list of verifiers
 /// against the quote structure.
 #[derive(Clone, Debug, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
@@ -1399,429 +1216,6 @@ mod test {
             .debug(true)
             .verify(&get_ias_report())
             .expect("Could not verify IAS report");
-    }
-
-    /// Ensure an OK result with the expected MRENCLAVE value succeeds.
-    #[test]
-    fn mrenclave_ok() {
-        let verifier = MrEnclaveVerifier {
-            mr_enclave: MrEnclave::from(&MR_ENCLAVE),
-            config_ids: vec![],
-            sw_ids: vec![],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_OK.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(verifier.verify(&data))
-    }
-
-    /// Ensure an OK result with the wrong MRENCLAVE value fails.
-    #[test]
-    fn mrenclave_fail() {
-        let verifier = MrEnclaveVerifier {
-            mr_enclave: MrEnclave::from(&MR_SIGNER),
-            config_ids: vec![],
-            sw_ids: vec![],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_OK.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(!verifier.verify(&data))
-    }
-
-    /// Ensure a CONFIGURATION_NEEDED result with the expected MRENCLAVE and
-    /// allowed advisory passes.
-    #[test]
-    fn mrenclave_config_pass() {
-        let verifier = MrEnclaveVerifier {
-            mr_enclave: MrEnclave::from(&MR_ENCLAVE),
-            config_ids: vec!["INTEL-SA-00239".to_owned()],
-            sw_ids: vec!["INTEL-SA-00123".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(verifier.verify(&data))
-    }
-
-    /// Ensure a CONFIGURATION_NEEDED result with the expected MRENCLAVE but
-    /// unexpected advisory fails.
-    #[test]
-    fn mrenclave_config_fail() {
-        let verifier = MrEnclaveVerifier {
-            mr_enclave: MrEnclave::from(&MR_ENCLAVE),
-            config_ids: vec!["INTEL-SA-00123".to_owned()],
-            sw_ids: vec!["INTEL-SA-00239".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(!verifier.verify(&data))
-    }
-
-    /// Ensure a SW_HARDENING_NEEDED result with the expected MRENCLAVE and
-    /// advisory passes.
-    #[test]
-    fn mrenclave_sw_pass() {
-        let verifier = MrEnclaveVerifier {
-            mr_enclave: MrEnclave::from(&MR_ENCLAVE),
-            config_ids: vec!["INTEL-SA-00123".to_owned()],
-            sw_ids: vec!["INTEL-SA-00239".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_SW.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(verifier.verify(&data))
-    }
-
-    /// Ensure a SW_HARDENING_NEEDED result with the expected MRENCLAVE but
-    /// unexpected advisory fails.
-    #[test]
-    fn mrenclave_sw_fail() {
-        let verifier = MrEnclaveVerifier {
-            mr_enclave: MrEnclave::from(&MR_ENCLAVE),
-            config_ids: vec!["INTEL-SA-00239".to_owned()],
-            sw_ids: vec!["INTEL-SA-00123".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_SW.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(!verifier.verify(&data))
-    }
-
-    /// Ensure a CONFIGURATION_AND_SW_HARDENING_NEEDED result with the expected
-    /// MRENCLAVE and config advisory passes.
-    #[test]
-    fn mrenclave_config_sw_pass_config() {
-        let verifier = MrEnclaveVerifier {
-            mr_enclave: MrEnclave::from(&MR_ENCLAVE),
-            config_ids: vec!["INTEL-SA-00239".to_owned()],
-            sw_ids: vec!["INTEL-SA-00123".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG_SW.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(verifier.verify(&data))
-    }
-
-    /// Ensure a CONFIGURATION_AND_SW_HARDENING_NEEDED result with the expected
-    /// MRENCLAVE and hardening advisory passes.
-    #[test]
-    fn mrenclave_config_sw_pass_sw() {
-        let verifier = MrEnclaveVerifier {
-            mr_enclave: MrEnclave::from(&MR_ENCLAVE),
-            config_ids: vec!["INTEL-SA-00123".to_owned()],
-            sw_ids: vec!["INTEL-SA-00239".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG_SW.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(verifier.verify(&data))
-    }
-
-    /// Ensure a CONFIGURATION_AND_SW_HARDENING_NEEDED result with the expected
-    /// MRENCLAVE but an unexpected advisory fails.
-    #[test]
-    fn mrenclave_config_sw_fail() {
-        let verifier = MrEnclaveVerifier {
-            mr_enclave: MrEnclave::from(&MR_ENCLAVE),
-            config_ids: vec!["INTEL-SA-00123".to_owned()],
-            sw_ids: vec!["INTEL-SA-00123".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG_SW.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(!verifier.verify(&data))
-    }
-
-    /// Ensure an OK result with the expected MRSIGNER, product, and minimum
-    /// version passes.
-    #[test]
-    fn mrsigner_ok() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_SIGNER),
-            product_id: 0,
-            minimum_svn: 0,
-            config_ids: vec![],
-            sw_ids: vec![],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_OK.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(verifier.verify(&data))
-    }
-
-    /// Ensure an OK result with the expected MRSIGNER, product, and minimum
-    /// version passes.
-    #[test]
-    fn mrsigner_fail_notok() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_SIGNER),
-            product_id: 0,
-            minimum_svn: 0,
-            config_ids: vec![],
-            sw_ids: vec![],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(!verifier.verify(&data))
-    }
-
-    /// Ensure an OK result with an unexpected MRSIGNER fails.
-    #[test]
-    fn mrsigner_fail_mrsigner() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_ENCLAVE),
-            product_id: 0,
-            minimum_svn: 0,
-            config_ids: vec![],
-            sw_ids: vec![],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_OK.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(!verifier.verify(&data))
-    }
-
-    /// Ensure an OK result with an unexpected product ID fails
-    #[test]
-    fn mrsigner_fail_product_id() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_SIGNER),
-            product_id: 1,
-            minimum_svn: 0,
-            config_ids: vec![],
-            sw_ids: vec![],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_OK.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(!verifier.verify(&data))
-    }
-
-    /// Ensure an OK result with a greater version fails
-    #[test]
-    fn mrsigner_fail_version() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_SIGNER),
-            product_id: 0,
-            minimum_svn: 1,
-            config_ids: vec![],
-            sw_ids: vec![],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_OK.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(!verifier.verify(&data))
-    }
-
-    /// Ensure a CONFIGURATION_NEEDED result with the expected MRSIGNER,
-    /// product, and minimum version passes, as long as the advisory is
-    /// accounted for
-    #[test]
-    fn mrsigner_pass_config() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_SIGNER),
-            product_id: 0,
-            minimum_svn: 0,
-            config_ids: vec!["INTEL-SA-00239".to_owned()],
-            sw_ids: vec![],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(verifier.verify(&data))
-    }
-
-    /// Ensure a SW_HARDENING_NEEDED result with the expected MRSIGNER,
-    /// product, and minimum version passes, as long as the advisory is
-    /// accounted for
-    #[test]
-    fn mrsigner_pass_sw() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_SIGNER),
-            product_id: 0,
-            minimum_svn: 0,
-            config_ids: vec![],
-            sw_ids: vec!["INTEL-SA-00239".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_SW.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(verifier.verify(&data))
-    }
-
-    /// Ensure a CONFIGURATION_AND_SW_HARDENING_NEEDED result with the expected
-    /// MRSIGNER, product, and minimum version passes, as long as the
-    /// advisory is accounted for.
-    #[test]
-    fn mrsigner_pass_sw_config_via_sw() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_SIGNER),
-            product_id: 0,
-            minimum_svn: 0,
-            config_ids: vec![],
-            sw_ids: vec!["INTEL-SA-00239".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG_SW.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(verifier.verify(&data))
-    }
-
-    /// Ensure a CONFIGURATION_AND_SW_HARDENING_NEEDED result with the expected
-    /// MRSIGNER, product, and minimum version passes, as long as the
-    /// advisory is accounted for.
-    #[test]
-    fn mrsigner_pass_sw_config_via_config() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_SIGNER),
-            product_id: 0,
-            minimum_svn: 0,
-            config_ids: vec![],
-            sw_ids: vec!["INTEL-SA-00239".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG_SW.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(verifier.verify(&data))
-    }
-
-    /// Ensure a CONFIGURATION_AND_SW_HARDENING_NEEDED result with the expected
-    /// MRSIGNER, and minimum version, but the wrong product fails, even if
-    /// the advisory is accounted for.
-    #[test]
-    fn mrsigner_fail_sw_config_for_product() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_SIGNER),
-            product_id: 1,
-            minimum_svn: 0,
-            config_ids: vec![],
-            sw_ids: vec!["INTEL-SA-00239".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG_SW.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(!verifier.verify(&data))
-    }
-
-    /// Ensure a CONFIGURATION_AND_SW_HARDENING_NEEDED result with the expected
-    /// MRSIGNER and product, but an earlier version, fails, even if the
-    /// advisory is accounted for.
-    #[test]
-    fn mrsigner_fail_sw_config_for_version() {
-        let verifier = MrSignerVerifier {
-            mr_signer: MrSigner::from(&MR_SIGNER),
-            product_id: 0,
-            minimum_svn: 1,
-            config_ids: vec![],
-            sw_ids: vec!["INTEL-SA-00239".to_owned()],
-        };
-
-        let report = VerificationReport {
-            sig: Default::default(),
-            chain: vec![],
-            http_body: IAS_CONFIG_SW.trim().to_owned(),
-        };
-
-        let data = VerificationReportData::try_from(&report).expect("Could not parse IAS result");
-        assert!(!verifier.verify(&data))
     }
 
     /// When the quote contains the basename we're expecting
