+++ attest/verifier/src/quote.rs
@@ -1,0 +1,4 @@
+// Copyright (c) 2018-2021 The MobileCoin Foundation
+
+//! Verifiers which operate on contents of the
+//! [`Quote`](::mc_attest_core::Quote) data structure.
