+++ attest/verifier/src/report_body.rs
@@ -1,0 +1,4 @@
+// Copyright (c) 2018-2021 The MobileCoin Foundation
+
+//! Verifiers which operate on the [`ReportBody`](::mc_attest_core::ReportBody)
+//! data structure.
